requirements.txt
Django==4.2.15
django-cors-headers==3.13.0
requests==2.31.0
urllib3
whitenoise==6.2.0
psycopg2-binary==2.9.3
dj-database-url==1.0
django-simple-history
